Unified logic for stable and normal farm cards in PoolCardTri
