+++ src/pages/EarnTri/EarnTriStable.tsx
@@ -1,13 +1,13 @@
 import React from 'react'
 import { AutoColumn } from '../../components/Column'
-import StablePoolCardTRI from '../../components/earn/StablePoolCardTri'
+import PoolCardTRI from '../../components/earn/PoolCardTri'
 import { FarmTabs } from '../../components/NavigationTabs'
 import { useFarms } from '../../state/stake/apr'
 import { TYPE } from '../../theme'
 import { isTokenAmountPositive } from '../../utils/pools'
 import { PoolSection, DataRow } from './EarnTri.styles'
 
-const MemoizedStablePoolCardTRI = React.memo(StablePoolCardTRI)
+const MemoizedPoolCardTRI = React.memo(PoolCardTRI)
 
 export default function EarnTri({ stablePoolsOrder }: { stablePoolsOrder: number[] }) {
   const allFarmArrs = useFarms()
@@ -27,7 +27,7 @@ export default function EarnTri({ stablePoolsOrder }: { stablePoolsOrder: number
         <PoolSection>
           {stablePools.map(farm =>
             farm.stableSwapPoolName == null ? null : (
-              <MemoizedStablePoolCardTRI
+              <MemoizedPoolCardTRI
                 key={farm.ID}
                 apr={farm.apr}
                 apr2={farm.apr2}
